+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_ after 2-1.
 Changed:
 
 - Removed local file logging
+- Add pymalcolm version tag to Block meta
 
 `4-0b1`_ - 2019-05-03
 ---------------------